NightAtTheMuseum
This was the fifth project in Udacity's VR Developer nanodegree. The purpose of this project was to construct a virtual reality exhibit about a cutting-edge VR company or technologies. After doing research, I have created a mobile virtual reality experience with “information booths” that include both visual and audio feedback for users. These booths will need to showcase my findings—the goal. this app going to inform other people about the topic at hand in a fun and creative way through a series of (at least five) display points. Users should be able to travel back and forth between display areas inside the space. Note: This is not a VR slideshow, it is a VR space which contains information about my research. This project showcases locomotion, VR scene design, interactivity, as well as an informed understanding of the industry.